Force permanent: true, use the module for sources.
When the offline option is used, permanent must be set to
true. This was the reason adding/removing sources and
interfaces was not working before and the "command"
module was used. Unfortunately, using the firewalld module
for adding/removing interfaces throws an exception. Until I
figure out what the issue is, keep using the command module.